Rename base64_util to base64_utilities to avoid file name conflicts. Courtesy of @JKoss2
