+++ src/base64_utilities.c
@@ -1,4 +1,4 @@
-#include "base64_util.h"
+#include "base64_utilities.h"
 
 static unsigned char base64_chars[] = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
